requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
groq==0.11.0
python-dotenv==1.0.0
httpx==0.27.2